hide particles when user prefers reduced motion
An implementation of hiding the floaty particles.
a sprinkle of js added to the default page template.

- Media query to determine if user prefers reduced-motion.
- Toggle the particles based on user preference.
Run script on page load, and add listener to detect if the preference changes.

Im not sure where the 'partical' js came from.
Would happily intertact with the particals directly through a config/api if someone could point me in the direction of where 'assets/particles/hadron.d57b1908.js' came from.