Fully remove particles (#17)
# Reason for Change
The floaty particles that make up the background of the Ruby Central
ghost theme are not universally appreciated. This change fully removes
the particles.

## Description of Change
This is very similar to the changes proposed in
https://github.com/rubycentral/rubycentral-theme/pull/15.

The only difference between the two changes is that this also removes
the 'hideParticles' Javascript that I added on #16.